--- src/ggomnbd_PAlive.h
@@ -1,33 +1,0 @@
-#ifndef GGOMNBD_PALIVE_HPP
-#define GGOMNBD_PALIVE_HPP
-arma::vec ggomnbd_PAlive(const double r,
-                         const double b,
-                         const double s,
-                         const arma::vec& vX,
-                         const arma::vec& vT_x,
-                         const arma::vec& vT_cal,
-                         const arma::vec& vAlpha_i,
-                         const arma::vec& vBeta_i);
-
-arma::vec ggomnbd_staticcov_PAlive(const double r,
-                                   const double alpha_0,
-                                   const double b,
-                                   const double s,
-                                   const double beta_0,
-                                   const arma::vec& vX,
-                                   const arma::vec& vT_x,
-                                   const arma::vec& vT_cal,
-                                   const arma::vec& vCovParams_trans,
-                                   const arma::vec& vCovParams_life,
-                                   const arma::mat& mCov_life,
-                                   const arma::mat& mCov_trans);
-
-arma::vec ggomnbd_nocov_PAlive(const double r,
-                               const double alpha_0,
-                               const double b,
-                               const double s,
-                               const double beta_0,
-                               const arma::vec& vX,
-                               const arma::vec& vT_x,
-                               const arma::vec& vT_cal);
-#endif
